refactor: rename mapNFTItems > mapNFTItem
